test: modularize unit and integration test suite for equalize operation
